Task 14 corrected

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -9,7 +9,7 @@
  */
 int main(void)
 {
-	long long i, j, n;
+	long i, j, n;
 	int k;
 
 	i = 1;
